+++ src/interfaces/derivatives.ts
@@ -1,0 +1,10 @@
+export type Derivatives = Record<string, DerivativesValue>;
+
+export interface DerivativesValue {
+  type: DerivativesType;
+  options?: unknown;
+  isNullable?: boolean;
+  nullablePercentage?: number;
+}
+
+export type DerivativesType = "string-interpolation" | "copy" | "date-before" | "date-after";